requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.4.2
python-multipart>=0.0.6
pandas>=2.1.0
pandasai>=3.0.0b2
python-dotenv>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0 